% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escreve.R
\name{escreve}
\alias{escreve}
\title{Eco}
\usage{
escreve(msg)
}
\arguments{
\item{msg}{(string) mensagem a ser repetida}
}
\value{
(string) mensagem repetida
}
\description{
Repete a mensagem passada.
}
\examples{

escreve("aloalo")

}
