% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soma.R
\name{get_content}
\alias{get_content}
\title{Get Content}
\usage{
get_content(url)
}
\arguments{
\item{url}{(string) url para o site que queremos pegar o content}
}
\value{
(html_document)
}
\description{
Get request + content
}
\examples{

get_content("http://google.com")

}
