% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soma.R
\name{soma}
\alias{soma}
\title{Soma}
\usage{
soma(a, b)
}
\arguments{
\item{a}{(numeric) Um dos números para somar}

\item{b}{(numeric) O outro dos números para somar}
}
\value{
(numeric) Número resultante da soma de a + b
}
\description{
Função que soma dois números.
}
\examples{

soma(3, 10)
soma(-2, 2)

}
